@@ -1,6 +1,8 @@
+// Using tailwind css v3
+// https://tailwindcss.com/docs/upgrade-guide
+
 module.exports = {
-	mode: 'jit',
-	purge: ['./index.html', './src/**/*.{js,ts,jsx,tsx}'],
+	content: ['./index.html', './src/**/*.{js,ts,jsx,tsx}'],
 	theme: {
 		extend: {
 			colors: {
@@ -16,6 +18,5 @@ module.exports = {
 			},
 		},
 	},
-	variants: {},
 	plugins: [],
 };
